+++ project-conventions/src/test/kotlin/io/github/jdbenitez94/criollo/kmp/foundation/conventions/ConventionFilesTest.kt
@@ -1,6 +1,6 @@
 package io.github.jdbenitez94.criollo.kmp.foundation.conventions
 
-import kotlin.test.Test
+import org.junit.jupiter.api.Test
 import kotlin.test.assertContains
 import kotlin.test.assertTrue