+++ runtime/src/jvmTest/kotlin/io/github/jdbenitez94/criollo/kmp/foundation/runtime/RetryWithBackoffTest.kt
@@ -1,12 +1,12 @@
 package io.github.jdbenitez94.criollo.kmp.foundation.runtime
 
 import kotlinx.coroutines.test.runTest
+import org.junit.jupiter.api.Test
 import strikt.api.expectThat
 import strikt.api.expectThrows
 import strikt.assertions.isEqualTo
 import kotlin.coroutines.cancellation.CancellationException
 import kotlin.random.Random
-import kotlin.test.Test
 import kotlin.time.Duration.Companion.milliseconds
 
 class RetryWithBackoffTest {